project/frontend/requirements.txt
streamlit
pandas
requests
plotly
pillow 